+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fixed : remove stat tab menu border on dark mode
 - fixed : remove mouse cursor when datatables not clickable
 - fixed : tab nb item update from child component with store
+- fixed : move worker logic in js file
 
 ## 0.8.17 (2024-10-20)